+++ compiler-core/src/type_/pattern.rs
@@ -610,7 +610,7 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                 ..
             } => match name.as_str() {
                 "true" | "false" => {
-                    self.error(Error::LowercaseBooleanPattern { location });
+                    self.error(Error::LowercaseBoolPattern { location });
                     Pattern::Invalid { location, type_ }
                 }
                 _ => {