+++ CHANGELOG.md
@@ -12,7 +12,7 @@
   of custom type variants.
   ([Adi Salimgereyev](https://github.com/abs0luty))
 
-- The "Did you want a Bool" error is no longer a syntax error, but instead a
+- The lowercase bool pattern error is no longer a syntax error, but instead a
   part of the analysis step. This allows the entire module to be analyzed, rather
   than stopping at the syntax error.
   ([mxtthias](https://github.com/mxtthias))